refactor(frame-studio): Refactor frame fucntionality; Implement clear button for slected image and share button for generated blob
