+++ src/app/frame/constants.ts
@@ -1,0 +1,11 @@
+export const MIN_ZOOM = 0;
+export const MAX_ZOOM = 5;
+
+// Available frames - frame0 to frame4
+export const AVAILABLE_FRAMES = [
+  "/frames/frame0.svg",
+  "/frames/frame1.svg",
+  "/frames/frame2.svg",
+  "/frames/frame3.svg",
+  "/frames/frame4.svg",
+];
