+++ src/app/frame/types.ts
@@ -1,0 +1,11 @@
+export type FitMode = "fit" | "fill";
+
+export type Point = {
+  x: number;
+  y: number;
+};
+
+export type ImageSize = {
+  width: number;
+  height: number;
+};
