+++ src/app/frame/utils/clamp.ts
@@ -1,0 +1,2 @@
+export const clamp = (value: number, min: number, max: number) =>
+  Math.min(Math.max(value, min), max);
