+++ src/app/frame/utils/share.ts
@@ -1,0 +1,4 @@
+const SHARE_TEXT =
+  "I'm attending DevFest 2025! 🚀 Join me for an amazing day of learning and networking. #DevFest2025 #GDGBaroda";
+
+export const getShareText = () => SHARE_TEXT;
