+++ src/components/Background.tsx
@@ -3,17 +3,34 @@
 import React, { memo, useEffect, useState } from "react";
 import { AuroraBackground } from "./ui/aurora-background";
 
-const Background = memo(() => {
-  const [isMounted, setIsMounted] = useState(false);
+type BackgroundProps = {
+  deferUntilIdle?: boolean;
+};
+
+const Background = memo(({ deferUntilIdle = false }: BackgroundProps) => {
+  const [shouldRenderAurora, setShouldRenderAurora] = useState(!deferUntilIdle);
 
   useEffect(() => {
-    setIsMounted(true);
-  }, []);
+    if (!deferUntilIdle) {
+      setShouldRenderAurora(true);
+      return;
+    }
 
-  if (!isMounted) {
+    let cancelled = false;
+    const timeout = setTimeout(() => {
+      if (!cancelled) setShouldRenderAurora(true);
+    }, 250);
+
+    return () => {
+      cancelled = true;
+      clearTimeout(timeout);
+    };
+  }, [deferUntilIdle]);
+
+  if (!shouldRenderAurora) {
     return (
       <div className="fixed inset-0 -z-10 pointer-events-none bg-zinc-50">
-        <div className="w-full h-full" />
+        <div className="h-full w-full" />
       </div>
     );
   }
