+++ src/components/ClientLayout.tsx
@@ -1,6 +1,7 @@
 "use client";
 
 import dynamic from "next/dynamic";
+import { usePathname } from "next/navigation";
 
 const Background = dynamic(() => import("@/components/Background"), {
   ssr: false,
@@ -15,10 +16,13 @@ export default function ClientLayout({
 }: {
   children: React.ReactNode;
 }) {
+  const pathname = usePathname();
+  const deferBackground = pathname?.startsWith("/frame");
+
   return (
     <>
       <SmoothScroll />
-      <Background />
+      <Background deferUntilIdle={Boolean(deferBackground)} />
       <div className="flex-1 flex flex-col">{children}</div>
     </>
   );
